installer: warn user about two conflicting installers
Since we might use signed EFI blobs we can't use random
unique PARTUUIDs for rootfs on installation media. If the
target hardware has got more than one slot for installation
media it's possible to plug in two installation media
with the same PARTUUID of their rootfs partitions.
This may result in installing OS from a wrong partition.

This change helps to detect and give warnings if both the SD
and USB slots have installation media with the same PARTUUID
plugged in.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>