meta-ostro-xt: swupd-client: add ostroproject.org public key
swupd-client 3.x checks the update server's certificate by comparing
its public key to what is stored on the device.

By default the public key is what is stored in swupd-client's GIT
repo. However, that is not the public key from ostroproject.org and
thus all communication to it fails.

To fix this, install the correct public key that belongs to
ostroproject.org.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>