systemd: freeze on Power Key press
Configure logind to suspend upon Power Key press. The desired (and
the only possible on this platform) suspend state is 'freeze'. Be
explicit about that setting by adding it in sleep.conf.

An additional change is needed in systemd-logind's capabilities.

systemd-logind checks whether /sys/power/state is writable but
gets 'no' because the SMACK label it runs (System) is not allowed
to write to floor label the file has. The process runs with a
reduced capability set but we additionally need CAP_MAC_OVERRIDE to
bypass SMACK MAC enforment.

The actual write to /sys/power/state happens in systemd-sleep
(from systemd-suspend.service) that runs unlimited capabilities.

Fixes: IOTOS-1758

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>