mesa: update to 12.0.2
A new version of mesa is needed to pull in the PCI IDs for i965/bxt
"2x6 variant".

https://lists.freedesktop.org/archives/mesa-commit/2016-May/064866.html

mesa.inc and disable-asm-on-non-gcc.patch are copied from the latest
openembedded core. mesa_12.0.2.bb is the same as what is being
proposed in https://patchwork.openembedded.org/patch/131361/

This commit shall be reverted once 12.0.2 is pulled in via ostro-os.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>