joule-udev-rules: enable power management for all USB devices
The Linux default is to keep USB devices powered on because there are
(broken) USB devices which don't work properly when power management
is active (see Documentation/usb/power-management.txt).

For Joule, the decision was made that this risk is acceptable
considering the power saving that can be obtained when enabling
USB power management.

Blacklist files (one for OS, one for the user) can be used to exclude
devices that do not work properly when autosuspend is enabled.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>